fix(api): isolate node-only auth codegen from browser typecheck
Extract BetterAuthType into auth.client-types.ts (only better-auth imports,
no node-only deps) so that web-spa can import it without pulling
auth-db.adapter.ts and auth-schema-codegen.ts into its TS compilation.

auth-db.adapter.ts retains the static import of auth-schema-codegen — it is
no longer in web-spa's transitive import graph.

Also ignore .boilerplate/ and .claude/ directories in ESLint config.
